03- SQL Server Connectivity with Python/requirements.txt
pymssql==2.3.2
python-dotenv==1.0.1
